// src/app/api/models/user-theme-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ThemeAccentDto } from '../models/theme-accent-dto';
import { ThemeDto } from '../models/theme-dto';

/**
 * DTO for user theme preferences
 */
export interface UserThemeDto {

  /**
   * The dark accent variant details
   */
  darkAccent?: ThemeAccentDto | null;

  /**
   * ID of the selected accent variant for dark mode
   */
  darkAccentId?: string;

  /**
   * The dark base theme details
   */
  darkTheme?: ThemeDto | null;

  /**
   * ID of the selected base theme for dark mode
   */
  darkThemeId?: string;

  /**
   * The light accent variant details
   */
  lightAccent?: ThemeAccentDto | null;

  /**
   * ID of the selected accent variant for light mode
   */
  lightAccentId?: string;

  /**
   * The light base theme details
   */
  lightTheme?: ThemeDto | null;

  /**
   * ID of the selected base theme for light mode
   */
  lightThemeId?: string;

  /**
   * Whether to sync with system settings
   */
  syncWithSystem?: boolean;
}